==> calmbelt/Step1_split_blast.ipynb <==
import glob

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

args = pickle.load(open('arguments.pickle', 'rb'))

from Bio import SeqIO

def batch_iterator(iterator, batch_size):
    entry = True  # Make sure we loop once

    while entry:
        batch = []

        while len(batch) < batch_size:
            try:
                entry = next(iterator)
            except StopIteration:
                entry = False

            if not entry:
                # End of file
                break

            batch.append(entry)

        if batch:
            yield batch


align_dir = os.path.join(args.genome_dir, 'align')
if not os.path.exists(align_dir):
    os.mkdir(align_dir)

data_list = []
n=0
for name in glob.glob(os.path.join(args.genome_dir, '*.fasta')):
    
    record_iter = SeqIO.parse(name, 'fasta')
    for i, batch in enumerate(batch_iterator(record_iter, 500), start=1):
        filename = os.path.join(align_dir, f'group{n}_{i}.fasta')
        count = SeqIO.write(batch, filename, 'fasta')
        print('Wrote {} records to {}'.format(count, filename))
    n+=1

# Blast

group_list = []
for name in glob.glob(os.path.join(align_dir, 'group*')):
    group_list += [name]

group_list

n_list = len(group_list)

n_list

import fnmatch, shutil

import subprocess
ref_path = args.reference_fname

for i, name in enumerate(group_list):   
    out_name = os.path.join(align_dir, f'alignment_{i}.xml')
    command_list = ['blastn', '-query', ref_path, '-subject', f'{name}', '-outfmt', '5', '-out', f'{out_name}']
    proc = subprocess.Popen(command_list, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
    proc.communicate()
