==> calmbelt/Step2_align_delete_insertion.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

from lib import utils

from Bio.Blast import NCBIXML

args = pickle.load(open('arguments.pickle', 'rb'))
align_dir = os.path.join(args.genome_dir, 'align')

# read alignment

alignment_list = []

for name in glob.glob(os.path.join(align_dir, 'alignment*.xml')):
    alignment_list += [name]

len(alignment_list)

snps_result = []
start_end_result = []

for name in alignment_list:
    result, start_end = utils.start_end_list_func(name)
    snps_result += result
    start_end_result += start_end


#start_end ของทุก alignment (1 sample อาจมีหลายค่า)
start_end_df = pd.DataFrame(start_end_result, columns=['id', 'start_end', 'length'])

id_list = sorted(list(set(start_end_df['id'])))
len(id_list)

start_end_df = start_end_df.sort_values(['id'])
start_end_df.head()

# delete insertion 

snps_df = pd.DataFrame(snps_result, columns=['id','query','position','sbjct', 'start_end', 'length'])
snps_df = snps_df.sort_values(['id','length','position'])

#save before delete insertion
snps_df.to_csv(os.path.join(args.preprocess_dir, 'snps_insertion_df.csv'), index=False)

#id ที่จะ delete
id_del = sorted(list(set(snps_df[snps_df['query']=='-']['id'])))
len(id_del)

#check ว่่าควรจะเหลือข้อมูลกี่แถว
snps_df.shape[0] -  snps_df[(snps_df['query']=='-')].shape[0]

#delete insertion
snps_df = utils.del_insertion(id_del, snps_df)

snps_df.shape

snps_df['change']=snps_df['query']+snps_df['position'].astype(str)+snps_df['sbjct']
snps_df = snps_df.sort_values(['id'])
snps_df.head()

snps_df.to_csv(os.path.join(args.preprocess_dir, 'snps_df.csv'), index=False)

start_end_df.to_csv(os.path.join(args.preprocess_dir, 'startstop_df.csv'), index=False)

#check ว่าไม่มีแล้วจริงๆ
snps_df[snps_df['query']=='-']
