==> calmbelt/Step3.1_insertion_position.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline

import warnings

warnings.filterwarnings('ignore')

from lib import utils,built

args = pickle.load(open('arguments.pickle', 'rb'))

# Load reference

seq_ref_list = pickle.load(open(os.path.join(args.preprocess_dir, 'seq_ref_list.pickle'), 'rb'))
seq_ref_length = len(seq_ref_list)
seq_ref_length

# หาตำแหน่งแต่ละ alignment ที่ใช้

startstop_df = pd.read_csv(os.path.join(args.preprocess_dir, 'startstop_df.csv'), index_col='id')
snps_df = pd.read_csv(os.path.join(args.preprocess_dir, 'snps_df.csv'))

#เรียง id
id_list = sorted(list(set(startstop_df.index)))
len(id_list)

#จำนวน sample ที่่ไม่่มี snps
len(set(id_list)-(set(list(snps_df['id']))))

#สร้าง array ขนาด sample x position จากนั้นใส่่ 1 ในตำแหน่่งที่มี
align_array = np.zeros((len(id_list), seq_ref_length), dtype=str)
snps_max_df = snps_df.copy()

sum_length,snps_max_df,align_array,choose_pos_list = utils.length_max_snps2(align_array,id_list,startstop_df,snps_max_df)

choose_pos_df = pd.DataFrame(choose_pos_list,columns=['id','start_new','end_new','start_end'])
choose_pos_df.head()

# หาตำแหน่ง insertion  และ จำนวน

snps_insertion_df = pd.read_csv(os.path.join(args.preprocess_dir, 'snps_insertion_df.csv'))

snps_insertion_df = snps_insertion_df[(snps_insertion_df['query']=='-')]

snps_insertion_df.head()

id_insertion = list(set(snps_insertion_df['id']))
len(id_insertion)

insertion_list = built.find_insertion_position(id_insertion,snps_insertion_df)

insertion_df = pd.DataFrame(insertion_list, columns=['id','position_ref','n_insert_base','start_end'])
insertion_df = insertion_df.sort_values('id')
insertion_df.head()

# นำมารวมกัน
- ตำแหน่งที่มี insertion กับ ตำแหน่งในสายที่ใช้สร้าง align array

#เอามารวมกันแล้ว จำนวนอาจลดลง เพราะบาง indertion อยู่ในสายที่เราไม่ได้ใช้ในการสร้าง align array
insertion_sum_df = pd.merge(insertion_df, choose_pos_df, how="right", on=["id", "start_end"])

insertion_sum_df = insertion_sum_df[insertion_sum_df['position_ref'].notna()]

insertion_sum_df = insertion_sum_df[(insertion_sum_df['position_ref']>=insertion_sum_df['start_new']) & (insertion_sum_df['position_ref']<=insertion_sum_df['end_new'])]

insertion_sum_df.head()

insertion_sum_df.to_csv(os.path.join(args.preprocess_dir, 'insertion_df.csv'), index=False)

insertion_sum_df
