==> calmbelt/Step3.2_create_insertion_array.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

from lib import utils

args = pickle.load(open('arguments.pickle', 'rb'))

# Load reference

seq_ref_list = pickle.load(open(os.path.join(args.preprocess_dir, 'seq_ref_list.pickle'), 'rb'))

# สร้าง array สำหรับ insertion

insertion_df = pd.read_csv(os.path.join(args.preprocess_dir, 'insertion_df.csv'))
startstop_df = pd.read_csv(os.path.join(args.preprocess_dir, 'startstop_df.csv'), index_col='id')

#เรียง id
id_list = sorted(list(set(startstop_df.index)))
len(id_list)

insertion_df

max_insertion = insertion_df.groupby(['position_ref']).max()

max_insertion = max_insertion.reset_index()
max_insertion.head()

max_insertion['sum'] = np.cumsum(max_insertion['n_insert_base'])
max_insertion = max_insertion[['position_ref','sum']]
max_insertion.head()

n_insertion = list(max_insertion['sum'])[-1]

#สร้าง array ขนาด sample x insertion จากนั้นใส่่ 1 ในตำแหน่่งที่มี
insertion_array = np.zeros((len(id_list),int(n_insertion)), dtype=str)

insertion_array.shape

s = 0
for p in list(max_insertion['position_ref']):
    temp_df = insertion_df[insertion_df['position_ref']==p]
    list_temp = list(temp_df['id'])

    for sample in list_temp:
        start = s
#             print(start)
        num = int(temp_df[temp_df['id']==sample]['n_insert_base'])
        i = id_list.index(sample)
        print(i)
        insertion_array[i][start:start+num] = 1
        print(p,insertion_array[i])
    s = int(max_insertion[max_insertion['position_ref']==p]['sum'].values)
    print(s)

insertion_array[insertion_array==''] = '0'

insertion_array

pickle.dump(insertion_array, open(os.path.join(args.preprocess_dir, 'insertion_array.pickle'), 'wb'))
