==> calmbelt/Step6_prepare_X_cluster.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

# สร้าง X_df

args = pickle.load(open('arguments.pickle', 'rb'))

max_etp_pos = pickle.load(open(os.path.join(args.preprocess_dir, 'max_etp_pos.pickle'), 'rb'))

align_array_df = pickle.load(open(os.path.join(args.preprocess_dir, "align_array_df.pickle"), "rb"))

# align_drop_df = pickle.load(open("../preprocessed/{}_align_drop_df.pickle".format(sample_type), "rb"))

align_array_df.head()

#เปลี่ยนชื่อคอลัมภ์เป็น str
align_array_df.columns = align_array_df.columns.astype(str)
align_array_df = align_array_df.sort_index()

X_df = align_array_df[max_etp_pos]
X_df = X_df.replace('','N')

X_df.shape

X_id_list = X_df.index
data = X_df.to_numpy()

pickle.dump(data, open(os.path.join(args.preprocess_dir, 'data_X_df.pickle'), 'wb'))
pickle.dump(X_id_list, open(os.path.join(args.preprocess_dir, 'X_id_list.pickle'), 'wb'))

# Sampling data

if args.subsampling_n_samples == -1:
    n = X_df.shape[0]
else:
    n = np.max([500, int(args.subsampling_n_samples)])

np.random.seed(8)
rand_pos = np.random.choice(range(X_df.shape[0]), n, replace=False)
len(set(rand_pos))

# rand_pos = sorted(rand_pos)

data_sampling = data[rand_pos]
data_sampling

id_sampling = X_id_list[rand_pos]

pickle.dump(rand_pos, open(os.path.join(args.preprocess_dir, 'rand_pos.pickle'), 'wb'))
pickle.dump(id_sampling, open(os.path.join(args.preprocess_dir, 'id_sampling.pickle'), 'wb'))
