==> calmbelt/Step7_pairwise_distance.ipynb <==
import sys
from scipy.spatial.distance import hamming

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

args = pickle.load(open('arguments.pickle', 'rb'))

# Load data

data = pickle.load(open(os.path.join(args.preprocess_dir, "data_X_df.pickle"), "rb"))

data

data.shape

n_sample = data.shape[0]
n_sample

# Create methods

import ray
num_cpus = os.cpu_count()

ray.init(num_cpus=num_cpus, ignore_reinit_error=True)

@ray.remote
class FeatureMatrix(object):
    def __init__(self, X):
        self.X = X
    def get_X(self):
        return self.X
    
    @ray.method(num_returns=1)
    def cal_dist(self, i):
        return np.array([hamming(self.X[i], x) for x in self.X[i:]], dtype=float)

FeatureMatrix.options(name="FeatureMatrix", lifetime="detached").remote(data)

X_actor = ray.get_actor("FeatureMatrix")
print (X_actor)
X_dat = ray.get(X_actor.get_X.remote())
print (X_dat.shape)

# Create Ray parallel actors = num_cpus

X_actor_list = []
for i in range(num_cpus):
    FeatureMatrix.options(name="FeatureMatrix_{}".format(i), lifetime="detached").remote(data)
    X_actor = ray.get_actor("FeatureMatrix_{}".format(i))
    X_actor_list += [X_actor]

X_actor_list

ray.get(X_actor_list[0].cal_dist.remote(n_sample-1))

# List and execute tasks

todo_tasks = [X_actor_list[i%num_cpus].cal_dist.remote(i) for i in range(n_sample)]
len(todo_tasks)

out = ray.get(todo_tasks)

# Create distance matrix

dist_mat = np.zeros((n_sample, n_sample))
iu = np.triu_indices(n_sample)
# il = np.tril_indices(n_sample)

dist_mat[iu] = np.hstack(out)

dist_mat

dist_mat = dist_mat + dist_mat.T
dist_mat.shape

dist_mat

pickle.dump(dist_mat, open(os.path.join(args.preprocess_dir, 'dist_mat_X_df.pickle'), 'wb'), protocol=4)


# Shutdown Ray

ray.shutdown()

```bash
jupyter nbconvert --ExecutePreprocessor.timeout=-1 --to notebook --execute "sandbox_pairwise_distance.ipynb"
```
