==> calmbelt/Step8_clustering.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

# โหลด distance matrix + clustering

args = pickle.load(open('arguments.pickle', 'rb'))

dm = pickle.load(open(os.path.join(args.preprocess_dir, "dist_mat_X_df.pickle"), "rb"))
dm

dm.shape

clade_df = pd.read_csv(os.path.join(args.preprocess_dir, 'clade_df.csv')).sort_values('Accession ID')
clade_df.head()

clade_df['Clade'] = clade_df['Clade'].replace(np.nan,'O')

from sklearn.cluster import KMeans
kmeans = KMeans(n_clusters=clade_df['WHO name'].unique().shape[0], 
                random_state=0, precompute_distances=False).fit(dm)

clade_df.loc[:, 'k_mean'] = [str(i) for i in kmeans.labels_]

clade_df.to_csv(os.path.join(args.preprocess_dir, 'clade_kmean.csv'), index=False)

X_id_list = pickle.load(open(os.path.join(args.preprocess_dir, "X_id_list.pickle"), "rb"))
rand_pos = pickle.load(open(os.path.join(args.preprocess_dir, "rand_pos.pickle"), "rb"))
rand_pos = sorted(rand_pos)

id_sampling = X_id_list[rand_pos]

clade_df = clade_df.set_index('Accession ID').loc[id_sampling].reset_index()
clade_df.columns = ['Accession ID'] + list(clade_df.columns[1:])
clade_df.head()

## TSE

# หา TSE เฉพาะ sampling data
dm_sampling = dm[rand_pos][:,rand_pos]
dm_sampling

dm_sampling.shape

from sklearn.manifold import TSNE
model = TSNE(metric="precomputed", n_components=2, perplexity=10.0, learning_rate=5.0, random_state=0)  #10,1
Y = model.fit_transform(dm_sampling) 

clade_df.head()

Y_df = pd.DataFrame(Y,columns=['x','y'])
Y_df['k_mean'] = clade_df['k_mean']
Y_df['result'] = clade_df['Clade']
Y_df['lineage'] = clade_df['Lineage']
Y_df['who_name'] = clade_df['WHO name']

Y_df.head()

pickle.dump(Y_df, open(os.path.join(args.preprocess_dir, 'Y_tse.pickle'), 'wb'))

# from lib import plot_data
# fig_k_mean = plot_data.fig_cluster(os.path.join(args.preprocess_dir, 'Y_tse.pickle'), 'k_mean')

# # fig_k_mean.write_image("../pic/fig_k_mean.png", scale=3)
# fig_k_mean.show()
