==> calmbelt/Step9_change_protein.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

from lib import utils

args = pickle.load(open('arguments.pickle', 'rb'))

# position each protein

# !pip install bcbio-gff 

# import pprint
# from BCBio.GFF import GFFExaminer
# from BCBio import GFF

# gfffile = "../COVID_data/GCF_009858895.2_ASM985889v3_genomic.gff"

# position_gene = []

# featureid_locations={}
# limits=dict(gff_type=["gene","mRNA","CDS"])
# with open(gfffile) as in_handle:
#     for rec in GFF.parse(in_handle, limit_info=limits):
#         for feature in rec.features:
#             print(feature.qualifiers['Name'],feature.location.start,feature.location.end)
#             position_gene += [[str(feature.qualifiers['Name'][0]),feature.location.start,feature.location.end]]
# #             print(feature.qualifiers['Name'])
# #             print(feature.qualifiers)


gene_data = pd.read_csv(args.reference_gene_loci, sep='\t')

gene_df = pd.DataFrame(gene_data)
gene_df.head()


gene_df.loc[:,'start']-=1
#start ที่่ 0

gene_df.head()

gene_df.to_csv(os.path.join(args.preprocess_dir, 'position_gene_df.csv'), index=False) 

# check amino change

seq_ref_list = pickle.load(open(os.path.join(args.preprocess_dir, "seq_ref_list.pickle"), "rb"))
align_array_df = pickle.load(open(os.path.join(args.preprocess_dir, 'align_array_df.pickle'), 'rb'))

snp_max_df = pd.read_csv(os.path.join(args.preprocess_dir, 'snps_max_df.csv'))

# snp_max_df = snp_max_df[snp_max_df['sbjct']!='-']
snp_max_df = snp_max_df[snp_max_df['sbjct']!='N']
snp_max_df = snp_max_df.sort_values('id')

snp_max_df = snp_max_df.reset_index()

change_df = snp_max_df.copy()
change_df = change_df[['id','position','query','sbjct']]
change_df['position'] -= 1 #ลบเพราะจะเทียบกับ ref ซึ่่งเริ่มที่ 0
change_df['check']='None'

change_protein_df = utils.change_protein_table(change_df,gene_df,align_array_df,seq_ref_list)
change_protein_df = change_protein_df[change_protein_df['check']==False].drop_duplicates(keep='first')

change_protein_df.head()

#change position & name gene NSP12
change_protein_df.loc[change_protein_df['gene']=='NSP12_2','position_protein']+=9
change_protein_df.loc[change_protein_df['gene']=='NSP12_2','gene']= 'NSP12'
change_protein_df.loc[change_protein_df['gene']=='NSP12_1','gene']= 'NSP12'

change_protein_df['change_protein'] = [str(q)+str(p)+str(s) for q,p,s in zip(change_protein_df['query_protein'],change_protein_df['position_protein'],change_protein_df['sbjct_protein'])]

change_protein_df.to_csv(os.path.join(args.preprocess_dir, 'summary_change_protein_df.csv'), index=False) 

change_protein_df=pd.read_csv(os.path.join(args.preprocess_dir, 'summary_change_protein_df.csv')) 

change_protein_df
