==> calmbelt/Step10_dendro.ipynb <==
from plotly.figure_factory import create_dendrogram
import numpy as np
import pandas as pd
from scipy.spatial.distance import pdist
import os, pickle

from lib import built, plot_data,utils

# Load data

args = pickle.load(open('arguments.pickle', 'rb'))

clade_df = pd.read_csv(os.path.join(args.preprocess_dir, 'clade_kmean.csv'))
clade_df.shape

clade_df.loc[:, 'Collection date'] = pd.to_datetime(clade_df['Collection date'])
clade_df.head()

clade_df.loc[:, 'Day'] = clade_df['Collection date'].dt.strftime('%d')

clade_df.head()

## Load  X

data = pickle.load(open(os.path.join(args.preprocess_dir, "data_X_df.pickle"), "rb"))

X = data

X.shape

X = np.where(X=='A', 1, X)
X = np.where(X=='T', 2, X)
X = np.where(X=='G', 3, X)
X = np.where(X=='C', 4, X)
X = np.where(X=='N', 0, X)
X = np.where(X=='-', 0, X)
X = np.where(X=='Y', 0, X)
X = np.where(X=='R', 0, X)
X = np.where(X=='S', 0, X)
X = np.where(X=='K', 0, X)
X = np.where(X=='M', 0, X)
X = np.where(X=='W', 0, X)

len(X)

# Sampling data for dendro

rand_pos = pickle.load(open(os.path.join(args.preprocess_dir, "rand_pos.pickle"), "rb"))

clade_df = clade_df.loc[rand_pos]
X = X[rand_pos]
X_id_list = list(clade_df['Accession ID'])

clade_df = clade_df.sort_values(['Accession ID'])
clade_df = clade_df.reset_index()

X.shape, len(X_id_list), clade_df.shape

clade_df

clade_df.columns = ['index', 'id', 'clade', 'lineage', 'date', 
                    'year', 'month', 'year_month', 'who_name', 'k_mean', 'day']

pickle.dump(clade_df, open(os.path.join(args.preprocess_dir, 'clade_dendro.pickle'), 'wb'))

pickle.dump(X, open(os.path.join(args.preprocess_dir, 'X_dendro.pickle'), 'wb'))

pickle.dump(X_id_list, open(os.path.join(args.preprocess_dir, 'X_id_list_dendro.pickle'), 'wb'))

# dendro

summary_change_protein_df = pd.read_csv(os.path.join(args.preprocess_dir, 'summary_change_protein_df.csv')) 
summary_change_protein_df = summary_change_protein_df[summary_change_protein_df['id'].isin(X_id_list)][['id','change_protein','gene','check']]

summary_change_protein_df

snps_max_df = pd.read_csv(os.path.join(args.preprocess_dir, 'snps_max_df.csv'))
snps_max_df.head()

fig_dendro_kmean,_,_ = plot_data.fig_dendro(args.preprocess_dir, 'k_mean')

# fig_dendro_kmean

fig_dendro_who,_,_ = plot_data.fig_dendro(args.preprocess_dir, 'who_name')

# fig_dendro_who

# fig_dendro_kmean.write_image("../pic/fig_dendro_kmean.png", scale=3)
# fig_dendro_who.write_image("../pic/fig_dendro_who.png", scale=3)
