==> calmbelt/Step11_alarm_more_patient.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

args = pickle.load(open('arguments.pickle', 'rb'))

# Load data

n=0
for name in glob.glob(os.path.join(args.genome_dir, '*patient*.tsv')):
    df_temp = pd.read_csv(name,sep='\t')
    if n == 0:
        df = df_temp
    else:
        df = pd.concat([df,df_temp], axis=0)
    n+=1

data_df = df[df['Host']=='Human']
data_df.shape

data = data_df[['Accession ID', 'Clade', 'Lineage', 'Collection date']]
data

data.loc[:, 'Collection date'] = pd.to_datetime(data['Collection date'])

data.loc[:, 'Year'] = data['Collection date'].dt.strftime('%Y')
data.loc[:, 'Month'] = data['Collection date'].dt.strftime('%m')
data.loc[:, 'Year-Month'] = data['Collection date'].dt.strftime('%Y-%m')

data['Accession ID'] = [i.split('_')[2] for i in data['Accession ID']]
data.sort_values(['Accession ID'])

who_name_df = pd.read_csv(args.predefined_label, index_col=0, sep='\t')

who_name_dict = who_name_df.T.to_dict('list')

# ใส่่ who name
who_name_list = []
for k in data['Lineage']:
    if k in list(who_name_dict.keys()):
        who_name_list += who_name_dict[k]
    else:
        who_name_list += ['Others']

data['WHO name'] = who_name_list

data.to_csv(os.path.join(args.preprocess_dir, 'clade_df.csv'), index=False)

set(data['Year-Month'])

id_month = {}
for m in list(set(data['Year-Month'])):
    id_month[m] = data[data['Year-Month']==m]['Accession ID'].values

pickle.dump(id_month, open(os.path.join(args.preprocess_dir, 'id_month.pickle'), 'wb'))
id_month_key = sorted(list(id_month.keys()))

# change protein each month

summary_change_protein_df = pd.read_csv(os.path.join(args.preprocess_dir, 'summary_change_protein_df.csv')) 

gene_list = list(set(summary_change_protein_df['gene']))
gene_list

len(set(summary_change_protein_df['gene']))

protein_alarm = {}
for g in gene_list:
    change_protein_df = summary_change_protein_df.copy()
    change_protein_df = change_protein_df[change_protein_df['check']==False]
    change_protein_df = change_protein_df[change_protein_df['gene']==g]
    protein_m_df = {}
    i=0
    for m in id_month_key:
        protein_m_df[m] = change_protein_df[change_protein_df['id'].isin(id_month[m].astype(int))]
        protein_m_df[m] = protein_m_df[m][['id','position_protein','change_protein']]
        protein_m_df[m] = protein_m_df[m].groupby(['position_protein','change_protein']).count().sort_values(by='position_protein') 
        protein_m_df[m]['month'] = i+1
        protein_m_df[m]['gene'] = g
        protein_m_df[m]['total'] = len(id_month[m])

        if i==0:
            protein_alarm_df = protein_m_df[m]
        else:
            protein_alarm_df = pd.concat([protein_alarm_df,protein_m_df[m]])
        i+=1
    
    protein_alarm[g] = protein_alarm_df.reset_index()
    protein_alarm[g]['graph'] = [p + ' : '+ str(n) + ' ' for p,n in zip(protein_alarm[g]['change_protein'],protein_alarm[g]['id'])]

len(sorted(protein_alarm.keys()))

from lib import plot_data
fig = plot_data.alarm_plot(protein_alarm['S'],[614,484],'position_protein')

pickle.dump(protein_alarm, open(os.path.join(args.preprocess_dir, 'protein_alarm.pickle'), 'wb'))

# change neucleotide each month

snps_max_df = pd.read_csv(os.path.join(args.preprocess_dir, 'snps_max_df.csv'))
snps_max_df

neucleotide_df = snps_max_df[snps_max_df['sbjct']!='-']
neucleotide_df

neucleotide_m_df = {}
i=0
for m in id_month_key:
    print(m)
    neucleotide_m_df[m] = neucleotide_df[neucleotide_df['id'].isin(id_month[m].astype(int))]
    neucleotide_m_df[m] = neucleotide_m_df[m][['id','position','change']]
    neucleotide_m_df[m] = neucleotide_m_df[m].groupby(['position','change']).count().sort_values(by='position')   
    neucleotide_m_df[m]['month'] = i+1
    neucleotide_m_df[m]['total'] = len(id_month[m])
    if i==0:
        neucleotide_alarm_df = neucleotide_m_df[m]
    else:
        neucleotide_alarm_df = pd.concat([neucleotide_alarm_df,neucleotide_m_df[m]])
    i+=1
# 

neucleotide_alarm_df = neucleotide_alarm_df.reset_index()  
neucleotide_alarm_df['graph'] = [p + ' : '+ str(n) + ' ' for p,n in zip(neucleotide_alarm_df['change'],neucleotide_alarm_df['id'])]
neucleotide_alarm_df

neucleotide_alarm_df.to_csv(os.path.join(args.preprocess_dir, 'neucleotide_alarm_df.csv'), index=False) 

# deletion each month

deletion_df = snps_max_df[snps_max_df['sbjct']=='-']
deletion_df

del_m_df = {}
i=0
for m in id_month_key:
#     print(m)
    del_m_df[m] = deletion_df[deletion_df['id'].isin(id_month[m].astype(int))]
    del_m_df[m] = del_m_df[m][['id','position']]
    del_m_df[m] = del_m_df[m].groupby('position').count().sort_values(by='position')   
    del_m_df[m]['month'] = i+1
    del_m_df[m]['total'] = len(id_month[m])
    if i==0:
        del_alarm_df = del_m_df[m]
    else:
        del_alarm_df = pd.concat([del_alarm_df,del_m_df[m]])
    i+=1

del_alarm_df = del_alarm_df.reset_index()

del_alarm_df.to_csv(os.path.join(args.preprocess_dir, 'del_alarm_df.csv'), index=False) 

# insertion each month

insertion_df = pd.read_csv(os.path.join(args.preprocess_dir, 'insertion_df.csv'))

insertion_df

insert_m_df = {}
i=0
for m in id_month_key:
#     print(m)
    insert_m_df[m] = insertion_df[insertion_df['id'].isin(id_month[m].astype(int))]
    insert_m_df[m] = insert_m_df[m][['id','position_ref','n_insert_base']]
    insert_m_df[m] = insert_m_df[m].groupby(['position_ref','n_insert_base']).count().sort_values(by='position_ref')   
    insert_m_df[m]['month'] = i+1
    insert_m_df[m]['total'] = len(id_month[m])
    if i==0:
        insert_alarm_df = insert_m_df[m]
    else:
        insert_alarm_df = pd.concat([insert_alarm_df,insert_m_df[m]])
    i+=1

insert_alarm_df = insert_alarm_df.reset_index()

insert_alarm_df

insert_alarm_df.to_csv(os.path.join(args.preprocess_dir, 'insert_alarm_df.csv'),index=False)
