==> calmbelt/Step12_meta_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy
%matplotlib inline

import os, pickle, re, glob, time
import pandas as pd
import numpy as np
np.set_printoptions(precision=2)

import matplotlib.pyplot as plt
import matplotlib.patches as patches
import seaborn as sns
from scipy import stats
from collections import Counter

sns.set_style('ticks')
pd.set_option('precision', 2)
#np.set_printoptions('precision', 2)

%matplotlib inline
import warnings
warnings.filterwarnings('ignore')

# Load data

args = pickle.load(open('arguments.pickle', 'rb'))
metadata = pd.read_csv(args.world_metadata_fname, sep='\t')

metadata = metadata[metadata['Host']=='Human']
metadata.shape

metadata.columns

metadata.loc[:, 'Collection date'] = pd.to_datetime(metadata['Collection date'])
metadata = metadata[metadata['Collection date'] >= '2020-01-01'].sort_values('Collection date')

metadata = metadata[['Accession ID','Location','Pango lineage','Clade','Collection date']]
metadata.head()

metadata.loc[:, 'Collection date'] = pd.to_datetime(metadata['Collection date'])

metadata['Accession ID'] = [i.split('_')[2] for i in metadata['Accession ID']]

metadata.loc[:, 'Year'] = metadata['Collection date'].dt.strftime('%Y')
metadata.loc[:, 'Month'] = metadata['Collection date'].dt.strftime('%m')
metadata.loc[:, 'Year-Month'] = metadata['Collection date'].dt.strftime('%Y-%m')

metadata['region'] = [(i.split('/')[0]).strip() for i in metadata['Location']]
metadata['country'] = [(i.split('/')[1]).strip() for i in metadata['Location']]

month_list = sorted(list(set(metadata['Year-Month'])))
month_list

for i,m in enumerate(month_list):
    temp_df = metadata[metadata['Year-Month']==m]
    temp_df = temp_df.groupby(['country','Pango lineage']).count().sort_values(by='country') 

    temp_df = temp_df[['Collection date']]
    temp_df.columns = ['count']

    temp_df['country_1']=[x[0] for x in temp_df.index]
    temp_df['lin']=[x[1] for x in temp_df.index]
    
    sum_country = temp_df.groupby(['country_1']).sum()
    sum_country_dict = sum_country.to_dict()['count']

    temp_df['sum'] = [sum_country_dict[i] for i in list(temp_df['country_1'])]
    temp_df['percentage'] = temp_df['count']/temp_df['sum']*100    
    temp_df['month'] = str(m)
    
    if i==0:
        count_metadata_df = temp_df
    else:
        count_metadata_df = pd.concat([count_metadata_df,temp_df])

count_metadata_df

count_metadata_df.to_csv(os.path.join(args.preprocess_dir, 'count_metadata_df.csv'), index=False)

# count_metadata_df[(count_metadata_df['lin']=='B.1.617.2')].sort_values('country')

# count_metadata_df[(count_metadata_df['country_1']=='Thailand')]
